requirements.txt
python-dotenv
gradio
google-genai
selenium
webdriver-manager
openai